SOLR-15740: Convert v2 schema-edit APIs to annotations
Solr's been in the slow process of moving its v2 APIs away from the
existing apispec/mapping framework towards one that relies on more
explicit annotations to specify API properties.

This commit converts all of Solr's schema-edit APIs to a single
annotation-based endpoint.

Note that the details of this conversion cause the introspect output of
this API to be lost.