SOLR-16392:  Refactor v2 collection 'LIST' API
This commit modifies the existing "List Collection" API to use the new
JAX-RS framework.  This allows the code implementing the list-collection
functionality to be more strongly typed, and adds the API to Solr's
growing OpenAPI spec.

Additionally, this PR consolidates the endpoints exposing this
functionality.  The same functionality was previously exposed at both
`GET /api/collections` and `GET /api/clusters`.  This commit removes the
second of these bindings.

This is a breaking change for users of the v2 API, but one that is
allowed because of v2's "experimental" designation.