First attempt at startup benchmark
Currently fails under jmh.sh due to exhausting filehandles on my
machine.  I have a theory that this is related to an incorrect
assumption I made about how often SolrStartup.stopJettyServerIfNecessary
would be run, relative to the number of benchmark-method invocations.

See that method for some details in the comments.