WIP: Attempt to swap out Jersey for RESTEasy
Putting a pin in this temporarily to try a different approach that I can
run tests on overnight.

The RESTEasy swap does appear do-able.  I was able to convert
JerseyApplications over to using the 'Application' interface (instead of
ResourceConfig).  The dep changes seem to be relatively straightforward.

Where I'm currently hung up is in finding a replacement for the Jersey
ApplicationHandler class, which was a nice internal abstraction that
offered a relatively clean interface to the rest of Solr.  I haven't
quite found the code that'd replace this in RESTEasy.  That hasn't
stalled out per se, there's just a lot of code to understand in
HttpServletDispatcher, SynchronousDispatcher, etc.