SOLR-16615: Share Jersey apps across multiple cores
This commit introduces the first pieces needed to share Jersey
applications across cores which share a configset.  This is done by
moving ownership of the Jersey 'ApplicationHandler' from the SolrCore
object itself to a new 'JerseyAppHandlerCache' class.  Reference counts
are used to track how many SolrCores are currently relying on a
particular ApplicationHandler; ApplicationHandlers are removed from the
cache when this reference count hits 0.