SOLR-16880: Add OAS to 'assembleRelease' task
This adds our OAS to the release artifacts gathered by the
'assembleRelease' task in gradle.  It produces a checksum for the OAS
file and stores the file in a separate 'oas' directory (similar to the
'changes' directory which already exists for Changes.html).

One remaining piece here is signing - this commit attempts to produce an
'*.asc' signature for the OAS when signing is enabled, but doesn't quite
have it working yet.