WIP: Convert FileStore APIs to JAX-RS
This currently doesn't work, due it looks like to some serialization
issues.  The first thing to get working is the GET
/node/files/$Path?meta=true case, since it's used by other APIs.
Currently this API returns a pretty bad serialization of the POJO.

What needs done is to take the FileStore.FileDetails interface, and
create a POJO version(s) of it that can serialize the data.  I started
to do this, creating 'FileStoreMetadata' (as a base),
FileStoreFileMetadata, and FileStoreDirectoryMetadata (not yet created)
as POJOs in the 'api' module.  But I didn't get all the way through that
before I had to stop for the day.  There's some still untangling to be
done of what the API actually returns in each case (metadata for a
directory, metadata for a file, how are extra signatures specified, the
FileStore.Metadata impl seems to support user-specified props...are
those returned? etc.)

Anyway, this seems close to done now, but definitely still some work
involved.