SOLR-17687: Remove arbitrary-props from MODIFYCOLL
Feature is redundant with the better known and document "collection
properties" feature.  This commit removes support for this legacy
feature which was previously deprecated in 9.x.

This starts the ReindexCollectionCmd change, but still requires test
changes that haven't yet been tried.  This was complicated a bit at the
time because 'waitForState' is mostly based around DocCollection
watchers, which don't actually contain collection-properties.

There's a larger question though of whether this is the right direction
or not.  See SOLR-16305 discussion for more details.

Still TODO:
  - ReindexCmd modification (esp. tests)
  - check
  - test

THis st